requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
transformers
faiss-cpu
torch
pydantic